added Makefile

+++ run.sh
@@ -1,5 +1,12 @@
-EXECUTABLE="dka2mka"
+EXECUTABLE="dka-2-mka"
 SOURCE="dka2mka.hs FaModule.hs"
+ZIPFILE="flp-fun-xkubis13.zip"
+
+if [ $1 == "zip" ]; then
+	echo "generating zip file..."
+	zip $ZIPFILE $SOURCE "README"
+	exit
+fi
 
 if [ -f $EXECUTABLE ]; then
 	rm $EXECUTABLE
